requirements.txt
psycopg[binary]==3.3.2
psycopg-pool==3.3.0
reflex==0.8.26
openai==2.21.0
databricks-sdk==0.88.0
python-dotenv==1.2.1
